filecfg: do not print tlaw_bilinear offset
The offset value is homogenous to temp_t and it cannot be represented
as is in a meaningful way to the user. Don't print it for now.

If printing is required it needs to be "back processed" and resolved
from the slope. It's probably better to separate internal (runtime)
and user-defined settings for this structure as well and always compute
the offset