timer: allow negative jitter
Due to wall clock resolution and sleep jitter, the timer loop can fire
again with a time difference that may be (slightly) < to the target.
This causes the iteration to be ignored and pushed to the next tperiod.

This change allows jitter to be both positive and negative and should
fix this minor issue.