Dbus: After=time-set.target
Services responsible for setting the system clock (CLOCK_REALTIME) from
a local source (such as a maintained timestamp file or imprecise
real-time clock) should pull in this target and order themselves before
it. Services where approximate, roughly monotonic time is desired should
be ordered after this unit, but not pull it in.

This target does not provide the accuracy guarantees of time-sync.target,
however does not depend on remote clock sources to be reachable, i.e.
the target is typically not delayed by network problems and similar. Use
of this target is recommended for services where approximate clock
accuracy and rough monotonicity is desired but activation shall not be
delayed for possibly unreliable network communication.