Alarms: implement notifier
This commit implements a basic execvp() notifier system: the optional
notifier will be executed with a list of error messages passed as its
arguments.

This commit is rather hackish, it hijacks the defconfig parser and
alarms_online() to set the notifier file/path. Eventually the alarms
subsystem will become a first-class citizen.