heatsource/boiler: jacket return itegral at -50%
During startup after a prolonged period of inactivity, the return temp
limitation when used without a return mixing valve would result in a
critical signal shift of -100% for consumers, completely shutting them
down. When the heating circuit is a 3-way mixing one, this means that
the boiler return can spend a long time without seeing any rise in
temperature, time during which the circuits do not receive any heat.

Mitigate this situation by limiting the return shift to -50%