hcircuit_logic(): correctly enforce flooring
In some situations, notably a transition to a lower power mode while the
ambient temperature is still rising and still not up to the new mode
request, the output flooring will not trigger as it should.

This commit bases the output flooring decision on absolute mode changes,
not on ambient temperature requests changes, to fix this.