hcircuit: fix output flooring again
After reviewing previous changes and the shortcomings of the last
incarnation, we've come full circle to something that is very close to
what was initially implemented in 999815a; with the main difference
being that flooring is now triggered only once at runmode change.

This still does not cover the case where boost *ends* during TRANS_UP,
which should also be accounted for.