hcircuit_logic(): always compute ambient model
The current logic would fail to correctly register transitions in the
ambient model, and it is impossible to fix this without reinstalling the
circular loop that was broken in 87023c4
The easiest way out is to always compute the ambient temperature in the
ambient model, even though this means the reported ambient temperature
may never converge. Since that modelled temperature should only be used
internally anyway, and since computing in TRANS_NONE does not affect
other parts of the code, it should be ok to do so.