scheduler_update_schedule(): break if found, not if existing
Before this commit, if system times jumped after rwchcd was started and
already synced on a schedule, the next correct schedule would not get
picked up as scheduler_ent_past_today() would return false and the else
if case of the for loop would exit the search for next schedule.

Furthermore, update the schedule only once at end of search.