dhwt: always trigger electric charge if skip_untrip is set
Rationale: if a DHWT has an embedded thermostat, this thermostat is let
in full control of the charge process when the DHWT run mode permits it
(COMFORT, ECO).

Thus, in these modes and if the tank has a thermostat, as soon as the
electric operation is engaged the DHWT will be allowed to charge.

There is no point in doing temperature control on the trip point alone
since we disable untrip afterwards: delegate temp control entirely.

This also solves the corner case of the external electric-only DHWT that
supplements an integrated boiler tank: when the switchover to electric
happens, the charge must begin at once since the boiler tank temperature
does not reflect the actual temperature of the electric tank.